+++ changelog.md
@@ -1,11 +1,15 @@
 # Change Log
 All notable changes to this project will be documented in this file.
  
-## [1.0.1] - 2021-05-27
+## [1.1] - 2021-06-01
 
 ### Added
+- create db for local transactions
+- automatically create polling and checkout page
+- automatically insert mobile number to checkout page to improve user experience
  
 ### Changed
+- implement mechanism to update payment via callback instead of polling
  
 ### Fixed